prep for package
