+++ index.html
@@ -7,7 +7,7 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/main.ts"></script>
 
     <split-panel style="width: 100%;height: 500px;">
       <div slot="resize:foo1">resize</div>
