+++ package.json
@@ -1,6 +1,20 @@
 {
   "name": "@madronejs/split-panel",
   "version": "0.0.0",
+  "type": "module",
+  "files": ["dist"],
+  "exports": {
+    ".": {
+      "import": "./index.js",
+      "require": "./index.cjs",
+      "types": "./index.d.ts"
+    },
+    "./*": {
+      "import": "./*/index.js",
+      "require": "./*/index.cjs",
+      "types": "./*/index.d.ts"
+    }
+  },
   "scripts": {
     "dev": "vite",
     "test": "vitest",
