--- src/SplitPanelView/index.ts
@@ -1,2 +1,0 @@
-export { default as SplitPanel } from './SplitPanel';
-export { default as SplitPanelView } from './SplitPanelView';
