+++ src/core/SplitPanel.spec.ts
