+++ src/core/SplitPanel.ts
