+++ src/core/defs.ts
