+++ src/core/flatten.mock.ts
