+++ src/core/flatten.spec.ts
