+++ src/core/flatten.ts
