+++ src/core/index.ts
@@ -1,0 +1,3 @@
+export { default as SplitPanel } from './SplitPanel';
+export * from './resize';
+export * from './flatten';
