+++ src/core/rawDefinition.spec.ts
