+++ src/core/resize.ts
