+++ src/index.ts
@@ -1,0 +1,4 @@
+import './style.scss';
+
+export { default as SplitPanel } from '@/core/SplitPanel';
+export { default as SplitPanelView } from '@/render/webComponent';
